interpolation search in array and demo
